refactor: 카드 프리뷰 레이아웃 책임 변경
