+++ src/stories/App.stories.tsx
@@ -1,8 +1,8 @@
-import type { Meta, StoryObj } from "@storybook/react";
-import App from "../App";
+import type { Meta, StoryObj } from '@storybook/react';
+import App from '../App';
 
 const meta = {
-  title: "App",
+  title: 'App',
   component: App,
 } satisfies Meta<typeof App>;
