+++ src/styles/App.style.tsx
@@ -4,6 +4,12 @@ export const AppLayout = styled.div`
   padding: 77px 30px;
 `;
 
+export const CardPreviewBox = styled.div`
+  display: flex;
+  justify-content: center;
+  width: 100%;
+`;
+
 export const CardForm = styled.div`
   display: flex;
   flex-direction: column;
