+++ src/styles/CardPreview.style.tsx
@@ -1,11 +1,5 @@
 import styled from 'styled-components';
 
-export const CardPreviewLayout = styled.div`
-  display: flex;
-  justify-content: center;
-  width: 100%;
-`;
-
 export const Card = styled.div`
   width: 212px;
   height: 132px;
